# Improve performance of `Parsed` and `Compiled`