Avoid quadratic scanning and copying of the program during parsing and compiling. This has no impact on the runtime performance, but improves parsing and compiling the program to operations between 50% and 100%:

```
     Running benches/compiled.rs (target/release/deps/compiled-b28e92b2ee828461)
Gnuplot not found, using plotters backend
parsed::e               time:   [21.955 µs 21.968 µs 21.982 µs]
                        change: [-47.977% -47.835% -47.629%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe

compiled::e             time:   [2.0509 µs 2.0544 µs 2.0580 µs]
                        change: [-91.418% -91.372% -91.326%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

parsed::fib             time:   [3.4734 µs 3.4814 µs 3.4953 µs]
                        change: [-34.800% -34.618% -34.452%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

compiled::fib           time:   [629.87 ns 631.43 ns 633.44 ns]
                        change: [-88.369% -88.318% -88.270%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high severe

parsed::mandelbrot      time:   [312.52 µs 312.73 µs 312.96 µs]
                        change: [-39.936% -39.739% -39.569%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe

compiled::mandelbrot    time:   [13.055 µs 13.081 µs 13.107 µs]
                        change: [-91.494% -91.422% -91.352%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

     Running benches/dispatch.rs (target/release/deps/dispatch-f8d8605c035faa06)
Gnuplot not found, using plotters backend
quicksort               time:   [14.645 ms 14.652 ms 14.659 ms]
                        change: [-0.3271% -0.1493% +0.0031%] (p = 0.08 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe
```